+++ src/app/view/mission/MissionSummaryButtonsComponent.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Information, Play } from 'mdi-material-ui';
+import { Link } from 'react-router-dom';
 import withWidth from 'material-ui/utils/withWidth';
 import Button from 'material-ui/Button';
 
@@ -10,17 +11,19 @@ class MissionSummaryButtonsComponent extends Component {
 	render() {
 		return <div>
 			<Button
-				color="secondary"
+				style={{ color: "#d32f2f" }}
 				size={this.props.width === "xs" ? "small" : "medium"}
-				onClick={() => this.props.history.push('/mission/'+this.props.mid)}
+				component={Link}
+				to={'/mission/'+this.props.mid}
 			>
 				<Information />
 				{I18n.t("Details")}
 			</Button>
 			<Button
-				color="secondary"
+				style={{ color: "#d32f2f" }}
 				size={this.props.width === "xs" ? "small" : "medium"}
-				onClick={() => this.props.history.push('/mission/'+this.props.mid+'/review')}
+				component={Link}
+				to={'/mission/'+this.props.mid+'/review'}
 			>
 				<Play />
 				{I18n.t("Start")}