+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - In review page, map keeps zoom level and used tiles between two features review
+- Links for missions in mission list can now be right-clicked for opening in new tab
 
 ### Fixed
 - "Show more pictures" button is hidden for complex geometries